added flag toenable skipping of data processing

+++ run/train.py
@@ -35,13 +35,14 @@ def main(cfg: TrainConfig):
         train_file.name
         for train_file in (Path(cfg.dir.processed_dir) / "train").glob("*.pkl")
     ]
-    if len(train_data_files) > 0:
-        LOGGER.info("Removing Previously Processed Files")
-        for file in train_data_files:
-            os.remove(Path(cfg.dir.processed_dir) / "train" / file)
+    if cfg.refresh_processed_data:
+        if len(train_data_files) > 0:
+            LOGGER.info("Removing Previously Processed Files")
+            for file in train_data_files:
+                os.remove(Path(cfg.dir.processed_dir) / "train" / file)
 
-    LOGGER.info("Processing Data for Loading")
-    pre_process_for_training(cfg)
+        LOGGER.info("Processing Data for Loading")
+        pre_process_for_training(cfg)
 
     seed_everything(cfg.seed)
     # init lightning model
